+++ yash-env/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Env::errexit_is_applicable`
+- `System::shell_path`
 - `SystemEx::fd_is_pipe`
 - `SystemEx::set_blocking`
 - `impl From<trap::Condition> for stack::Frame`